pyproject.toml
[tool.poetry]
name = "jane-doe"
version = "0.1.0"
description = ""
authors = ["beatricedaniel <beatrice.yk.daniel@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.7"
python-docx = "^1.1.2"
presidio-analyzer = "^2.2.355"
pandas = "^2.2.3"
presidio-anonymizer = "^2.2.356"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
